requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Deprecated==1.2.14
dj-database-url==2.0.0
Django==4.2.3
django-on-heroku==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.59.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
oauth2client==4.1.3
packaging==23.1
protobuf==4.24.0rc1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pyparsing==3.1.0
Pyrebase4==4.7.1
python-jwt==4.0.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
wrapt==1.15.0
